Remove start and end dates from reports
This was introduced because Alma provides this in their solution for RSS
feeds, but with hindsight I think this should not be an essential
feature of a new book report. It doesn't connect that well with the idea
of a feed, and I think it's better to streamline the RSS feed by having
it always include a fixed number of newest items, and then rather add other
representations for non-streamy stuff like "new books this month" in
addition.